+++ components/main/MainRoom.tsx
@@ -6,7 +6,8 @@ import { useRef, useState } from 'react'
 import { useRouter } from 'next/router'
 
 import { GLTFResult } from '../../lib/types/mainTypes'
-import { randomDummyMemory } from '@/lib/constant/dummyMemoryList'
+import { MemoryType } from '@/pages/memory-list'
+// import { randomDummyMemory } from '@/lib/constant/dummyMemoryList'
 
 export default function MainRoom(props: JSX.IntrinsicElements['group']) {
   const { nodes, materials } = useGLTF('/main_room.gltf') as GLTFResult
@@ -75,6 +76,47 @@ export default function MainRoom(props: JSX.IntrinsicElements['group']) {
   })
 
   const routeToRandomCardModal = () => {
+    const dummyMemoryList: MemoryType[] = [
+      {
+        id: 50,
+        backgroundImage: 'https://res.cloudinary.com/djrrkk6k4/image/upload/v1687723452/cqqkj6yijqpi75ne3lba.jpg',
+        text: '기분내려고 신경 좀 썼지',
+        videoId: 'https://youtu.be/ucj8RMdkJ7Y',
+        createdAt: '2023-06-25 20:04:13',
+      },
+      {
+        id: 49,
+        backgroundImage: 'https://res.cloudinary.com/djrrkk6k4/image/upload/v1687723421/xs4rfjwcg92awtwkl041.jpg',
+        text: '일찍 일어난 보람이 있네',
+        videoId: 'https://youtu.be/ucj8RMdkJ7Y',
+        createdAt: '2023-06-25 20:03:42',
+      },
+      {
+        id: 48,
+        backgroundImage: 'https://res.cloudinary.com/djrrkk6k4/image/upload/v1687723399/mk0o3jkw9ht8flp4vtpc.jpg',
+        text: '학창시절의 책상, 추억이 새록새록 번져와',
+        videoId: 'https://youtu.be/ucj8RMdkJ7Y',
+        createdAt: '2023-06-25 20:03:19',
+      },
+      {
+        id: 46,
+        backgroundImage: 'https://res.cloudinary.com/djrrkk6k4/image/upload/v1687723312/pdtfyg3jjc86bv4ez8gp.jpg',
+        text: '나만 두고 가지마!',
+        videoId: 'https://youtu.be/ucj8RMdkJ7Y',
+        createdAt: '2023-06-25 20:01:53',
+      },
+      {
+        id: 44,
+        backgroundImage: 'https://res.cloudinary.com/djrrkk6k4/image/upload/v1687722937/hx6jafdpbbxz04xja5qm.jpg',
+        text: '오늘도 헤엄 헤엄',
+        videoId: 'https://youtu.be/ucj8RMdkJ7Y',
+        createdAt: '2023-06-25 19:55:38',
+      },
+    ]
+
+    const randomIndexOfDummyMemoryList = Math.floor(Math.random() * dummyMemoryList.length)
+    const randomDummyMemory = dummyMemoryList.filter((el, index) => index === randomIndexOfDummyMemoryList)[0]
+
     sessionStorage.setItem('dummyMemory', JSON.stringify(randomDummyMemory))
     router.push('/memory-list')
   }
